test: simplify structure
